pyproject.toml
[project]
name = "nexus-ai-team"
version = "0.1.0"
description = "NEXUS — Neural Executive Unified System"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "websockets>=14.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.28.0",
    "redis>=5.2.0",
    "psycopg[binary]>=3.2.0",
    "psycopg-pool>=3.2.0",
    "alembic>=1.14.0",
    "litellm>=1.75.0",
    "python-telegram-bot>=21.0",
    "PyYAML>=6.0",
    "aiohttp>=3.9.0",
    "psutil>=5.9.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.9.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.25.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"alembic/**" = ["E402"]
"tests/**" = ["E402"]
"heartbeat/service.py" = ["E402"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.setuptools]
packages = ["gateway", "nexus_v1", "pipeline", "db", "agents", "interfaces", "equipment", "company", "agentoffice", "qa", "heartbeat"]
